requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
psycopg2
requests
numpy
pandas
pillow